Do not clean article title with bleach
Titles can contain text or html so cleaning them before display is
necessary. However there is no need to do complex processing with them
(unlike the article body), so relying on Django for escaping is enough.

This fixes titles showing "foo &amp; bar" in the interface and is more
consistent with how article bodies work because we save titles
untouched in the database.