Allow to ingest feeds without published or updated date
Antirez's feed does not contain any date, thus the published_at is
modified to accept null values. When an article is ingested we still
have the created_at that we can use to sort articles.

This is why articles are now inserted in reverse order (usually newer
articles appear first in feeds).