Allow multiple users to add the same feed
The actual feed is now created in the background only if a valid RSS or
atom file can be found at the given URI.

The name of the feed is not taken from the user anymore but read from
the feed itself. Because of that the unique constraint on feed name is
lifted.