Relax X-Forwarded-For check
An error 400 used to be returned to clients presenting an XFF header
(like Opera Mini) because Caddy simply adds the remote IP to the
existing XFF header. The middleware in strict mode expects exactly 1
XFF IP.

The strict mode can be removed as the middleware correctly sets the
remote IP of the proxy even if the original request contains an XFF
header.